refactor wrappers
