+++ packages/waku/src/main.ts
@@ -1,4 +1,5 @@
-export { honoWrapper, connectWrapper } from './lib/middleware.js';
+export { honoWrapper } from './lib/middleware/honoWrapper.js';
+export { connectWrapper } from './lib/middleware/connectWrapper.js';
 export { rsc } from './lib/middleware/rsc.js';
 
 export async function build(options?: { ssr?: boolean }) {
